stake management refactor - fund view module
