+++ src/genesis.rs
@@ -8,6 +8,7 @@ use crate::node_activation::*;
 use crate::node_config::*;
 use crate::user_data::*;
 use crate::fund_transf_module::*;
+use crate::fund_view_module::*;
 use crate::user_stake::*;
 
 #[elrond_wasm_derive::module(GenesisModuleImpl)]
@@ -24,6 +25,9 @@ pub trait GenesisModule {
 
     #[module(FundTransformationsModuleImpl)]
     fn fund_transf_module(&self) -> FundTransformationsModuleImpl<T, BigInt, BigUint>;
+
+    #[module(FundViewModuleImpl)]
+    fn fund_view_module(&self) -> FundViewModuleImpl<T, BigInt, BigUint>;
 
     #[module(NodeConfigModuleImpl)]
     fn node_config(&self) -> NodeConfigModuleImpl<T, BigInt, BigUint>;
@@ -68,7 +72,7 @@ pub trait GenesisModule {
 
         // validate that node stake and user stake match
         let stake_required_by_nodes = BigUint::from(num_inactive_nodes) * self.node_config().get_stake_per_node();
-        let mut total_inactive_stake = self.user_data().get_user_stake_of_type(USER_STAKE_TOTALS_ID, UserStakeState::Inactive);
+        let mut total_inactive_stake = self.fund_view_module().get_user_stake_of_type(USER_STAKE_TOTALS_ID, UserStakeState::Inactive);
         if stake_required_by_nodes != total_inactive_stake {
             return sc_error!("stake required by nodes must match total user stake at genesis");
         }
