+++ src/unexpected.rs
@@ -6,6 +6,7 @@ use crate::settings::*;
 use crate::stake_sale::*;
 use crate::user_data::*;
 use crate::fund_transf_module::*;
+use crate::fund_view_module::*;
 
 imports!();
 
@@ -25,6 +26,9 @@ pub trait UnexpectedBalanceModule {
     #[module(FundTransformationsModuleImpl)]
     fn fund_transf_module(&self) -> FundTransformationsModuleImpl<T, BigInt, BigUint>;
 
+    #[module(FundViewModuleImpl)]
+    fn fund_view_module(&self) -> FundViewModuleImpl<T, BigInt, BigUint>;
+
     #[module(StakeSaleModuleImpl)]
     fn stake_sale(&self) -> StakeSaleModuleImpl<T, BigInt, BigUint>;
 
@@ -35,7 +39,7 @@ pub trait UnexpectedBalanceModule {
     /// This can come from someone accidentally sending ERD to the contract via direct transfer.
     #[view(getUnexpectedBalance)]
     fn get_unexpected_balance(&self) -> BigUint {
-        let mut expected_balance = self.user_data().all_funds_in_contract();
+        let mut expected_balance = self.fund_view_module().all_funds_in_contract();
         expected_balance += self.rewards().get_total_cumulated_rewards();
         expected_balance -= self.rewards().get_sent_rewards();
